Move many block methods to ArmiObject (temporarily).
Various block methods that do make logical sense at any level
of the reactor data model were moved to the ArmiObject. On one
hand this exasperates the fact that ArmiObject has way too
many public methods. On the other, it consolidates our tree-
traversing methods that are still in use into one place and
relieves some of the "specialness" of the Block. This is a
useful step in designing a better ArmiObject that will likely
rely more on composition of other functional classes which
will take these methods with them. A warning was added
to the code warning users that the ArmiObject API should
be considered quite unstable as we go through this.

* Remove the Batch class and its tests.

  The functionality of a Batch was useful but will be built in
  to the updated composite model once it's done. In the meantime
  it is not being used and is safe to remove.

  The motivation for removing it came when I noticed a discrepancy
  in the various set/add mass/masses methods.
